FORENSE/requirements.txt
Pillow
piexif
pytesseract
opencv-python
numpy
pandas
matplotlib
markdown
pytest
